Creating SubmissionCard component
